refactor(system): adjust files structure
